Minor change
git-svn-id: https://svn.ntop.org/svn/ntop/trunk/nDPI@8089 a02cd0c1-8e76-42e1-a119-56f9641475e2

+++ src/lib/protocols/ayiya.c
@@ -54,7 +54,7 @@ void ndpi_search_ayiya(struct ndpi_detection_module_struct *ndpi_struct, struct 
       u_int32_t fireyears = 86400 * 365 * 5;
       
 #ifndef __KERNEL__
-      now = time(NULL);
+      now = (u_int32_t)time(NULL);
 #else
       now = 1402729042; /* Dummy workaround */
 #endif
